+++ app/[locale]/projects/create-project/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useRef, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { ApolloError } from '@apollo/client';
 import { useRouter } from 'next/navigation';
 import { useTranslations } from 'next-intl';
@@ -30,6 +30,8 @@ import ErrorMessages from '@/components/ErrorMessages';
 
 //Other
 import logECS from '@/utils/clientLogger';
+import { scrollToTop } from '@/utils/general';
+import { routePath } from '@/utils/routes';
 import { useToast } from '@/context/ToastContext';
 
 interface CreateProjectResponse {
@@ -53,6 +55,7 @@ const ProjectsCreateProject = () => {
   const router = useRouter();
   //For scrolling to error in page
   const errorRef = useRef<HTMLDivElement | null>(null);
+  const inputFieldRef = useRef<HTMLInputElement | null>(null);
 
   const [fieldErrors, setFieldErrors] = useState<CreateProjectErrorsInterface>({
     projectName: '',
@@ -168,9 +171,6 @@ const ProjectsCreateProject = () => {
 
     // Update state with all errors
     setFieldErrors(errors);
-    if (errors) {
-      setErrors(Object.values(errors).filter((e) => e)); // Store only non-empty error messages
-    }
 
     return !hasError;
   };
@@ -190,9 +190,9 @@ const ProjectsCreateProject = () => {
       }
       setFormSubmitted(true)
     } catch (error) {
-      logECS('error', 'updateSection', {
+      logECS('error', 'createProject', {
         error,
-        url: { path: '/template/[templateId]/section/[sectionid]' }
+        url: { path: routePath('projects.create') }
       });
       if (error instanceof ApolloError) {
         setErrors(prevErrors => [...prevErrors, error.message]);
@@ -207,6 +207,7 @@ const ProjectsCreateProject = () => {
   const handleFormSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
 
+    setFormSubmitted(true);
     // Clear previous error messages
     setErrors([]);
 
@@ -228,6 +229,10 @@ const ProjectsCreateProject = () => {
       }
     }
   };
+
+  useEffect(() => {
+    scrollToTop(inputFieldRef);
+  },[fieldErrors.projectName])
 
   return (
     <>
@@ -253,12 +258,14 @@ const ProjectsCreateProject = () => {
           <ErrorMessages errors={errors} ref={errorRef} />
           <Form onSubmit={handleFormSubmit}>
             <FormInput
+              ref={inputFieldRef}
               name="projectName"
               type="text"
               value={formData.projectName}
               label={CreateProject('form.projectTitle')}
               helpMessage={CreateProject('form.projectTitleHelpText')}
-              isRequired={true}
+              isRequired={false}
+              ariaRequired={true}
               onChange={handleInputChange}
               isInvalid={(!formData.projectName || !!fieldErrors.projectName) && formSubmitted}
               errorMessage={fieldErrors.projectName.length > 0 ? fieldErrors.projectName : CreateProject('messages.errors.title')}
@@ -285,9 +292,8 @@ const ProjectsCreateProject = () => {
             <Button
               type="submit"
               className=""
-              onPress={() => { setFormSubmitted(true) }}
-            >{Global('buttons.continue')}
-
+            >
+              {Global('buttons.continue')}
             </Button>
 
           </Form>