+++ components/Form/FormInput/index.tsx
@@ -23,6 +23,7 @@ interface InputProps {
   inputClasses?: string;
   disabled?: boolean;
   isRequired?: boolean;
+  ariaRequired?: boolean;
   isInvalid?: boolean;
   errorMessage?: string;
   helpMessage?: string;
@@ -31,7 +32,7 @@ interface InputProps {
   pattern?: string;
 }
 
-const FormInput: React.FC<InputProps & React.InputHTMLAttributes<HTMLInputElement>> = ({
+const FormInput = React.forwardRef<HTMLInputElement, InputProps & React.InputHTMLAttributes<HTMLInputElement>>(({
   name,
   id,
   type,
@@ -47,6 +48,7 @@ const FormInput: React.FC<InputProps & React.InputHTMLAttributes<HTMLInputElemen
   inputClasses = '',
   disabled = false,
   isRequired = false,
+  ariaRequired = false,
   isInvalid = false,
   errorMessage = '',
   helpMessage = '',
@@ -54,7 +56,7 @@ const FormInput: React.FC<InputProps & React.InputHTMLAttributes<HTMLInputElemen
   maxLength = undefined,
   pattern,
   ...rest
-}) => {
+}, ref) => {
 
   return (
     <>
@@ -71,6 +73,7 @@ const FormInput: React.FC<InputProps & React.InputHTMLAttributes<HTMLInputElemen
           {description}
         </Text>
         <Input
+          ref={ref}
           id={id}
           name={name}
           type={type}
@@ -84,6 +87,7 @@ const FormInput: React.FC<InputProps & React.InputHTMLAttributes<HTMLInputElemen
           minLength={minLength}
           maxLength={maxLength}
           pattern={pattern}
+          aria-required={ariaRequired}
           {...rest}
         />
 
@@ -98,6 +102,7 @@ const FormInput: React.FC<InputProps & React.InputHTMLAttributes<HTMLInputElemen
       </TextField>
     </>
   );
-};
+});
 
+FormInput.displayName = 'FormInput';
 export default FormInput;